requirements.txt
# Core dependencies
streamlit==1.32.2
opencv-python==4.9.0.80
numpy==1.26.4
matplotlib==3.8.3
pillow==10.2.0
python-multipart==0.0.9
python-dotenv==1.0.1

# ML/AI dependencies
mediapipe==0.10.21
torch==2.5.1
transformers
timm

# Data processing
pandas==2.2.1
plotly==5.18.0
seaborn

# Video processing
moviepy==1.0.3
pytube==15.0.0

# 3D processing
open3d

# Monitoring and utilities
psutil>=5.9.0
rich>=10.14.0
python-jose==3.3.0

# Optional 3D dependencies (install manually)
# pip install "git+https://github.com/facebookresearch/pytorch3d.git"
# pip install kaolin==0.17.0 -f https://nvidia-kaolin.s3.us-east-2.amazonaws.com/torch-2.0.1_cu118.html
